staging/pkg/mod/golang.org/x/net@v0.0.0-20201110031124-69a78807bb2b/html/testdata/webkit/tests8.dat

#data
<div>
<div></div>
</span>x
#errors
(1,5): expected-doctype-but-got-start-tag
(3,7): unexpected-end-tag
(3,8): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "
"
|       <div>
|       "
x"

#data
<div>x<div></div>
</span>x
#errors
(1,5): expected-doctype-but-got-start-tag
(2,7): unexpected-end-tag
(2,8): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "x"
|       <div>
|       "
x"

#data
<div>x<div></div>x</span>x
#errors
(1,5): expected-doctype-but-got-start-tag
(1,25): unexpected-end-tag
(1,26): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "x"
|       <div>
|       "xx"

#data
<div>x<div></div>y</span>z
#errors
(1,5): expected-doctype-but-got-start-tag
(1,25): unexpected-end-tag
(1,26): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "x"
|       <div>
|       "yz"

#data
<table><div>x<div></div>x</span>x
#errors
(1,7): expected-doctype-but-got-start-tag
(1,12): foster-parenting-start-tag
(1,13): foster-parenting-character
(1,18): foster-parenting-start-tag
(1,24): foster-parenting-end-tag
(1,25): foster-parenting-start-tag
(1,32): foster-parenting-end-tag
(1,32): unexpected-end-tag
(1,33): foster-parenting-character
(1,33): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     <div>
|       "x"
|       <div>
|       "xx"
|     <table>

#data
<table><li><li></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <li>
|     <li>
|     <table>

#data
x<table>x
#errors
(1,1): expected-doctype-but-got-chars
(1,9): foster-parenting-character
(1,9): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     "xx"
|     <table>

#data
x<table><table>x
#errors
(1,1): expected-doctype-but-got-chars
(1,15): unexpected-start-tag-implies-end-tag
(1,16): foster-parenting-character
(1,16): eof-in-table
#document
| <html>
|   <head>
|   <body>
|     "x"
|     <table>
|     "x"
|     <table>

#data
<b>a<div></div><div></b>y
#errors
(1,3): expected-doctype-but-got-start-tag
(1,24): adoption-agency-1.3
(1,25): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <b>
|       "a"
|       <div>
|     <div>
|       <b>
|       "y"

#data
<a><div><p></a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,15): adoption-agency-1.3
(1,15): adoption-agency-1.3
(1,15): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|     <div>
|       <a>
|       <p>
|         <a>
